nixos/main.nix: Update options to use pkgs.frappix
Previously the options incorrectly only specified "frappix.foo", which
led to evaluation errors from missing attributes on frappix, depending on
what the frappix argument was. In particular in our system we had
"frappix" set to the input. Specifying pkgs.frappix.foo in full makes
this more resiliant to different configurations