Package wiki
Frappe wiki is self-described as a "Modern, Feature-Rich Wiki
Application" built with the frappe framework. We wanted to use it at
Clicks so we decided to package it for Frappix!

We had minor trouble with the pyproject.toml as it seems to be missing
some properties that made other things easy to package, however we found
ways to cajole nix into building it anyway