wiki: add missing metadata
Previously the wiki was missing its pname and version as we couldn't
extract them from the meta.

Unfortunately, some updates started using pname in a few places (e.g.
mkSiteAssets.nix) which broke any sites using wiki. By manually adding
the metadata we can sidestep this issue. It may be worth making an
upstream commit to add pyproject.toml so we can properly extract the
metadata, but that will have to be a followup sometime.